WIP: basics of msgpack encoding test
There are still some issues around py2/3 string handling.  The organization is improved but I'm not sure what the intent is with the other test subclasses; specifically would we want to run a test run with TestMessagesMongoDB+JSON and TestMessagesMongoDB+Msgpack?  Or is one subclass like the way I have it now sufficient?

Change-Id: Id35bf1dce9d290c0296f86477d6b59b214b9e951